+++ src/DiagramElement/Formatter/DiagramFormatter.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smeghead\PhpClassDiagram\DiagramElement\Formatter;
+
+interface DiagramFormatter {
+    public function head(): string;
+    public function tail(): string;
+}